+++ tests/golden/outputs/msl/6220-break-from-loop.metal
@@ -3,7 +3,7 @@ using namespace metal;
 
 void break_from_loop() {
     {
-        float local_0 = 0;
+        int local_0 = 0;
         while (true) {
             if ((local_0 < 4)) {
                 {
